requirements.txt
# Core dependencies
pydantic>=2.5.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
typer>=0.9.0
python-dotenv>=1.0.0

# LLM API
openai>=1.6.0
together>=1.0.0

# PDF processing
PyPDF2>=3.0.1

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0

# Google Scholar API
serpapi>=0.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
black>=23.12.0
mypy>=1.7.0

# Graph visualization
pyvis>=0.3.2
networkx>=3.0
numpy>=1.24,<2.0

